% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\name{msma_OneComp}
\alias{msma_OneComp}
\alias{normvec}
\alias{sparse}
\alias{cand4lambda}
\alias{searchseq}
\alias{project}
\alias{matserr}
\title{Internal functions}
\usage{
msma_OneComp(X, Y, Z = NULL, lambdaX = NULL, lambdaY = NULL,
  lambdaXsup = NULL, lambdaYsup = NULL, eta = 1, type = "lasso",
  inX = NULL, inY = NULL, inXsup = NULL, inYsup = NULL, muX = 0,
  muY = 0, dmode = "PLS", verbose = FALSE)

normvec(a)

sparse(x, lam, eta = 1, type = "lasso", inidx = NULL)

cand4lambda(x, len)

searchseq(orgseq, opt)

project(x, y)

matserr(X1, X2)
}
\description{
These are internal functions for \code{msma}
}
\details{
These are not intended for use by users.
}
\keyword{internal}
