% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{package}
\name{msma-package}
\alias{msma-package}
\title{Multiblock Sparse Multivariable Analysis Package}
\description{
A Package for implementation of multiblock multivariable data analysis.
}
\author{
Atsushi Kawaguchi. \email{kawa_a24@yahoo.co.jp}
}
\seealso{
\code{\link{msma}}
}
\keyword{documentation}

