% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{methods}
\name{strsimdata}
\alias{strsimdata}
\title{Structured Simulate Data sets}
\usage{
strsimdata(
  n = 100,
  ncomp = 5,
  WX = NULL,
  Xps = 10,
  Yps = FALSE,
  rho = 0.8,
  Z = FALSE,
  seed = 1,
  minpct = 0.25,
  maxpct = 0.75
)
}
\arguments{
\item{n}{a numeric scalar, sample size.}

\item{ncomp}{number of components}

\item{WX}{a matrix or a list, weights.}

\item{Xps}{a numeric vector, numbers of columns for X. The length of vector corresponds to the number of blocks.}

\item{Yps}{a numeric vector, numbers of columns for Y. The length of vector corresponds to the number of blocks.}

\item{rho}{a numeric, correlation}

\item{Z}{outcome}

\item{seed}{a seed number for generating random numbers.}

\item{minpct}{minimum percent of nonzero}

\item{maxpct}{maximum percent of nonzero}
}
\value{
\item{X}{Simulated X which has a list form}

\item{Y}{Simulated Y which has a list form}
}
\description{
This is a function for generating multiblock data based on the multivariable normal distribusion
}
\details{
The output is a list of matrics.
}
