% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\name{msma_OneComp}
\alias{msma_OneComp}
\alias{normvec}
\alias{sparse}
\alias{cand4lambda}
\alias{searchseq}
\alias{project}
\alias{matserr}
\alias{ginv}
\alias{hcmsma}
\alias{rmnorm}
\title{Internal functions}
\usage{
msma_OneComp(
  X,
  Y,
  Z = NULL,
  lambdaX = NULL,
  lambdaY = NULL,
  lambdaXsup = NULL,
  lambdaYsup = NULL,
  eta = 1,
  type = "lasso",
  inX = NULL,
  inY = NULL,
  inXsup = NULL,
  inYsup = NULL,
  muX = 0,
  muY = 0,
  dmode = "PLS",
  verbose = FALSE,
  intseed = 1
)

normvec(a)

sparse(x, lam, eta = 1, type = "lasso", inidx = NULL)

cand4lambda(x, len)

searchseq(orgseq, opt)

project(x, y)

matserr(X1, X2)

ginv(X, tol = sqrt(.Machine$double.eps))

hcmsma(
  object,
  nclust = 4,
  graph = FALSE,
  hmethod = "ward.D2",
  axes = c(1, 2),
  block = "block",
  XY = "X"
)

rmnorm(
  n,
  mean = rep(0, nrow(sigma)),
  sigma = diag(length(mean)),
  method = c("eigen", "svd", "chol"),
  pre0.9_9994 = FALSE,
  checkSymmetry = TRUE
)
}
\description{
These are internal functions for \code{msma}
}
\details{
These are not intended for use by users.
}
\keyword{internal}
