% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\name{msma_OneComp}
\alias{msma_OneComp}
\alias{normvec}
\alias{sparse}
\alias{cand4lambda}
\alias{searchseq}
\alias{project}
\alias{matserr}
\alias{strsimdata}
\alias{ginv}
\title{Internal functions}
\usage{
msma_OneComp(X, Y, Z = NULL, lambdaX = NULL, lambdaY = NULL, eta = 1,
  type = "lasso", inX = NULL, inY = NULL, muX = 0, muY = 0,
  dmode = "PLS", verbose = FALSE, intseed = 1)

normvec(a)

sparse(x, lam, eta = 1, type = "lasso", inidx = NULL)

cand4lambda(x, len)

searchseq(orgseq, opt)

project(x, y)

matserr(X1, X2)

strsimdata(n = 100, ncomp = 5, Xps = 10, Yps = FALSE, rho = 0.8,
  Z = FALSE, seed = 1, minpct = 0.25, maxpct = 0.75)

ginv(X, tol = sqrt(.Machine$double.eps))
}
\description{
These are internal functions for \code{msma}
}
\details{
These are not intended for use by users.
}
\keyword{internal}
