% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{package}
\name{msma-package}
\alias{msma-package}
\title{Multiblock Sparse Matrix Analysis Package}
\description{
A Package for Implementation of the method
}
\references{
Kawaguchi A, Yamashita F (2017). Supervised Multiblock Sparse Multivariable Analysis with Application to Multimodal Brain Imaging Genetics. Biostatistics, 18(4) 651-665.
}
\seealso{
\code{\link{msma}}
}
\author{
Atsushi Kawaguchi. \email{kawa_a24@yahoo.co.jp}
}
\keyword{documentation}
