% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{methods}
\name{predict.msma}
\alias{predict.msma}
\title{Prediction}
\usage{
\method{predict}{msma}(object, newX, newY = NULL, ...)
}
\arguments{
\item{object}{an object of class "\code{msma}." Usually, a result of a call to \code{msma}}

\item{newX}{a matrix in which to look for the variables used to predict X. This is required.}

\item{newY}{a matrix in which to look for the variables used to predict Y.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\item{X}{predicted X}

\item{sbX}{block score for X}

\item{Y}{predicted Y}

\item{sbY}{block score for Y}
}
\description{
predict method for class "msma".
}
\details{
This function produces a prediction from new data based on \code{\link{msma}} fit. It is mainly used in cross-validation
}
\examples{
##### data #####
tmpdata = simdata(n = 50, rho = 0.8, Yps = c(10, 12, 15), Xps = 20, seed=1)
X = tmpdata$X; Y = tmpdata$Y 

##### Two Component #####
fit2 = msma(X, Y, comp=2, lambdaX=2, lambdaY=1:3)
summary(fit2)

##### Predict #####
test = predict(fit2, newX=X, newY=Y)

}
