% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.r
\docType{methods}
\name{simdata}
\alias{simdata}
\title{Simulate Data sets}
\usage{
simdata(n = 100, rho = 0.8, Yps = c(100, 120, 150), Xps = 500, seed = 1)
}
\arguments{
\item{n}{a numeric scalar, sample size.}

\item{rho}{a numeric scalar, correlation coefficient.}

\item{Yps}{a numeric vector, numbers of columns for Y. The length of vector corresponds to the number of blocks.}

\item{Xps}{a numeric vector, numbers of columns for X. The length of vector corresponds to the number of blocks.}

\item{seed}{a seed number for generating random numbers.}
}
\value{
\item{X}{Simulated X which has a list form}

\item{Y}{Simulated Y which has a list form}
}
\description{
This is a function for generating multiblock data based on the multivariable normal distribusion
}
\details{
The output is a list of matrics.
}
